# src/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AI Tutor API",
    "description": "API for AI Tutor, managing students, courses, and AI-generated questions. For more details, visit the [GitHub repository](https://github.com/ShishirShekhar/E-Learning-Platform.git).",
    "version": "1.0.0",
    "contact": {
      "name": "Shishir Shekhar",
      "email": "sspdav02@gmail.com"
    },
    "termOfService": "https://www.google.com",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local server"
    }
  ],
...
  "paths": {
...
    "/api/v1/instructor/create-course": {
      "post": {
        "tags": [
          "Instructor"
        ],
        "summary": "Create a new course",
        "description": "Endpoint to create a new course",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Course information",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "type": "string",
                  "example": "Introduction to AI"
                },
                "description": {
                  "type": "string",
                  "example": "Learn the basics of AI"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Course created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "title": {
                          "type": "string",
                          "example": "Introduction to AI"
                        },
                        "description": {
                          "type": "string",
                          "example": "Learn the basics of AI"
                        }
                      }
                    },
                    "error": {}
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "title": {
                          "type": "string",
                          "example": "Introduction to AI"
                        },
                        "description": {
                          "type": "string",
                          "example": "Learn the basics of AI"
                        }
                      }
                    },
                    "error": {}
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Required fields are missing",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {},
                    "error": {
                      "type": "string",
                      "example": "Required fields are missing"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {},
                    "error": {
                      "type": "string",
                      "example": "Required fields are missing"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {},
                    "error": {
                      "type": "string",
                      "example": "Internal Server Error"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {},
                    "error": {
                      "type": "string",
                      "example": "Internal Server Error"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/instructor/create-test": {
      "post": {
        "tags": [
          "Instructor"
        ],
        "summary": "Create a new test with frequency",
        "description": "Endpoint to create a new test with frequency",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Test information",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "courseId": {
                  "type": "string",
                  "example": "60a8b6d5a3f1b1b0b4f9e1d2"
                },
                "topic": {
                  "type": "string",
                  "example": "Introduction to AI"
                },
                "frequency": {
                  "type": "string",
                  "example": "daily"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Test created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "topic": {
                          "type": "string",
                          "example": "Introduction to AI"
                        },
                        "frequency": {
                          "type": "string",
                          "example": "daily"
                        }
                      }
                    },
                    "error": {}
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "topic": {
                          "type": "string",
                          "example": "Introduction to AI"
                        },
                        "frequency": {
                          "type": "string",
                          "example": "daily"
                        }
                      }
                    },
                    "error": {}
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Required fields are missing",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {},
                    "error": {
                      "type": "string",
                      "example": "Required fields are missing"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {},
                    "error": {
                      "type": "string",
                      "example": "Required fields are missing"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Course not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {},
                    "error": {
                      "type": "string",
                      "example": "Course not found"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {},
                    "error": {
                      "type": "string",
                      "example": "Course not found"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {},
                    "error": {
                      "type": "string",
                      "example": "Internal Server Error"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {},
                    "error": {
                      "type": "string",
                      "example": "Internal Server Error"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}